+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # DeepGNN Changelog
 
 ## 0.1 DeepGNN
+
+* Refactor decoder/converter interface for speed. Add new base input format.